requirements.txt
flask~=3.0.2
python-dotenv
PyJWT
streamlit~=1.30.0
flask-restx
tensorflow~=2.15.0
werkzeug~=3.0.1
pillow~=10.2.0
flask-cors
pathlib~=1.0.1
matplotlib~=3.8.2
numpy~=1.26.3